refactor: changed name of utils file
